HpManager refactor
The HpManager folder has been replaced by the Optimizer folder. HpManager.py has been split into 3 differents files to split the different classes and their subclasses.
