+++ Optimizer/Optimizer.py
@@ -1,5 +1,5 @@
 """
-    @file:              HpManager.py
+    @file:              Optimizer.py
     @Author:            Alexandre Ayotte
     @Creation Date:     09/07/2020
 
@@ -10,41 +10,11 @@
 
 from abc import ABC, abstractmethod
 from copy import deepcopy
-from enum import Enum, unique
-from GPyOpt.methods import BayesianOptimization
-from hyperopt import hp, fmin, rand, tpe
-import numpy as np
 from functools import partial
-
-@unique
-class HPtype(Enum):
-    """
-    Class containing possible types of hyper-parameters
-
-    From @Ref1
-    """
-
-    real = 1
-    integer = 2
-    categorical = 3
-
-
-class Hyperparameter:
-    """
-    From @Ref1
-    """
-    def __init__(self, name, _type, value=None):
-        """
-        Class that defines an hyper-parameter
-
-        :param name: Name of the hyper-parameter
-        :param _type: One type out of HPtype (real,.integer, categorical)
-        :param value: List with the value of the hyper-parameter
-        """
-
-        self.name = name
-        self.type = _type
-        self.value = value
+from GPyOpt.methods import BayesianOptimization
+from hyperopt import fmin, rand, tpe
+import numpy as np
+from Optimizer.HpSpace import HyperoptSearchSpace, GPyOptSearchSpace
 
 
 def get_optimizer(hp_space, algo, n_rand_point=10, acquisition_fct="EI"):
@@ -262,190 +232,3 @@ class GpyOptOptimizer(HPOptimizer):
 
             hp_list = bo.suggest_next_locations(pending_X=pending_x)
             return self._list_to_dict(hp_list[0])
-
-
-class SearchSpace:
-    """
-    From @Ref1
-    """
-
-    def __init__(self, space):
-        """
-        Definition of a search space for our hyper-parameters
-        """
-
-        self.default_space = space
-        self.space = space
-        self.log_scaled_hyperparam = []
-
-    def reset(self):
-        """
-        Resets search space to default
-        """
-
-        self.space = deepcopy(self.default_space)
-        self.log_scaled_hyperparam.clear()
-
-    def __getitem__(self, key):
-        return self.space[key]
-
-    def __setitem__(self, key, value):
-        self.space[key] = value
-
-
-class HyperoptSearchSpace(SearchSpace):
-    """
-    From @Ref1
-    """
-    def __init__(self, hp_space):
-        """
-        Class that defines a compatible search space with Hyperopt package hyper-parameter optimization algorithm
-
-        :param hp_space: A dictionary that represent the hyperparameters space.
-        """
-        space = {}
-        self.discrete_space = {}
-
-        for hparam_name in sorted(hp_space.keys()):
-            space[hparam_name] = hp_space[hparam_name].compatible_format('tpe', hparam_name)
-
-            if isinstance(hp_space[hparam_name], DiscreteDomain):
-                self.discrete_space[hparam_name] = hp_space[hparam_name].values
-
-        super(HyperoptSearchSpace, self).__init__(space)
-
-    def get_discrete_index(self, hparam_names, value):
-        """
-        Return the index list of a value in the possible values list according to the name of the hyperparameter.
-
-        :param hparam_names: A string that represent the hyperparameter name.
-        :param value: The value. Must be present in the possible values list
-        :return: The index list of the value in the list.
-        """
-
-        return self.discrete_space[hparam_names].index(value)
-
-
-class GPyOptSearchSpace(SearchSpace):
-    """
-    From @Ref1
-    """
-    def __init__(self, hp_space):
-        """
-        Class that defines a compatible search space with GPyOpt package hyper-parameter optimization algorithm
-
-        :param hp_space: A dictionary that represent the hyperparameters space.
-        """
-
-        space = {}
-
-        for hparam_name in sorted(hp_space.keys()):
-            hparam = hp_space[hparam_name]
-
-            if isinstance(hparam, ContinuousDomain):
-                _type = 'continuous'
-            else:
-                _type = 'discrete'
-
-            space[hparam_name] = {'name': hparam_name,
-                                  'type': _type,
-                                  'domain': hparam.compatible_format('gaussian_process', hparam_name),
-                                  'dimensionality': 1}
-
-        super(GPyOptSearchSpace, self).__init__(space)
-
-
-@unique
-class DomainType(Enum):
-    """
-    Class containing possible types of hyper-parameters
-
-    From @Ref1
-    """
-
-    continuous = 1
-    discrete = 3
-
-
-class Domain:
-    """
-    From @Ref1
-    """
-    def __init__(self, type_):
-        """
-        Abstract (parent) class that represents a domain for hyper-parameter's possible values
-
-        :param type_: One type of domain among DomainType
-        """
-
-        self.type = type_
-
-
-class ContinuousDomain(Domain):
-    """
-    From @Ref1
-    """
-    def __init__(self, lower_bound, upper_bound):
-        """
-        Class that generates a continuous domain
-
-        :param lower_bound: Lowest possible value (included)
-        :param upper_bound: Highest possible value (included)
-        """
-
-        if lower_bound > upper_bound:
-            lower_bound, upper_bound = upper_bound, lower_bound
-
-        self.lb = lower_bound
-        self.ub = upper_bound
-
-        super(ContinuousDomain, self).__init__(DomainType.continuous)
-
-    def compatible_format(self, tuner_method, label):
-        """
-        Builds the correct format of a uniform distribution according to the method used by the tuner
-
-        :param tuner_method: Name of the method employed by the HPtuner.
-        :param label: String defining the name of the hyper-parameter
-        :return: Uniform distribution compatible with method used by HPtuner
-        """
-
-        if tuner_method in ['random_search', 'tpe', 'annealing']:
-            return hp.uniform(label, self.lb, self.ub)
-
-        elif tuner_method == 'gaussian_process':
-            return tuple([self.lb, self.ub])
-
-
-class DiscreteDomain(Domain):
-    """
-    Took from @Ref1
-    """
-    def __init__(self, possible_values):
-        """
-        Class that generates a domain with possible discrete values of an hyper-parameter
-
-        :param possible_values: list of values
-        """
-
-        self.values = possible_values
-
-        super(DiscreteDomain, self).__init__(DomainType.discrete)
-
-    def compatible_format(self, tuner_method, label):
-        """
-        Builds the correct format of discrete set of values according to the method used by the tuner
-
-        :param tuner_method: Name of the method employed by the HPtuner.
-        :param label: String defining the name of the hyper-parameter
-        :return: Set of values compatible with method used by HPtuner
-        """
-
-        if tuner_method == 'grid_search':
-            return self.values
-
-        elif tuner_method in ['random_search', 'tpe', 'annealing']:
-            return hp.choice(label, self.values)
-
-        elif tuner_method == 'gaussian_process':
-            return tuple(self.values)
